// src/app/interfaces/user.model.ts
export class User {
  id: number = 0;
  email: string = '';
  username: string = '';
  password: string = '';
  firstname: string = '';
  lastname: string = '';
  city: string = '';
  street: string = '';
  number: number = 0;
  zipcode: string = '';
  lat: string = '';
  long: string = '';
  phone: number = 0;
}